main: changes, fixes
- refactors http session usage across modules
- adjusts video info extraction
- improves temporary directory, concurrency limits, buffering, and error handling in download routines

+++ util/av/videoinfo.go
@@ -10,9 +10,9 @@ func GetVideoInfo(filePath string) (int64, int64, int64) {
 	if err != nil {
 		return 0, 0, 0
 	}
-	duration := gjson.Get(probeData, "format.duration").Int()
+	duration := gjson.Get(probeData, "format.duration").Float()
 	width := gjson.Get(probeData, "streams.0.width").Int()
 	height := gjson.Get(probeData, "streams.0.height").Int()
 
-	return duration, width, height
+	return int64(duration), width, height
 }
